Add synchronisation for ELKA reference entities to the ELKASync module
The ELKASync module now checks every sended entity for reference entities.
If one is attached to the parent entity, the reference is sended after the
successful transmission of the parent.
The only reference currently supported is: Enleitungsstelle->Klaerwerk.
Further references may need further updates of the ELKAsync module or the
reference mapping.